pyproject.toml
[project]
name = "auto-instagram"
version = "0.1.0"
description = "Open-source Instagram publisher that drives instagram.com with Playwright and imported cookies."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "auto-instagram contributors" }]
keywords = ["instagram", "automation", "playwright", "publishing", "scheduler"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
]

dependencies = [
    "patchright>=1.55.0",
    "typer>=0.15.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.2",
    "apscheduler>=3.10.4",
    "rich>=13.9.0",
    "tzdata>=2024.2",
]

[project.optional-dependencies]
camoufox = ["camoufox>=0.4.0"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
auto-ig = "auto_instagram.cli:app"

[project.urls]
Homepage = "https://github.com/auto-instagram/auto-instagram"
Issues = "https://github.com/auto-instagram/auto-instagram/issues"

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_instagram"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers"
markers = [
    "e2e: end-to-end tests that hit a real Instagram account (opt-in)",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
